fixup! evtimer: add msg helper functions
